chore: refactor iOS config reader
